refactor to remove docker builds from terraform

+++ terraform/deploy.sh
@@ -10,6 +10,16 @@ CAL_BASE_IMAGE="stsci/hst-pipeline:CALDP_20210505_CAL_final"
 # this is the tag that the image will have in AWS ECR
 CALDP_IMAGE_TAG="latest"
 
+# these variables are overrides for developers that allow the deploy script to build from local calcloud/caldp source
+# i.e. CALCLOUD_BUILD_DIR="$HOME/deployer/calcloud"
+# these can be set as environment variables before running to avoid changing the script directly
+# (and avoid accidentally committing a custom path to the repo...)
+CALCLOUD_BUILD_DIR=${CALCLOUD_BUILD_DIR:-""} 
+CALDP_BUILD_DIR=${CALDP_BUILD_DIR:-""}
+
+#uncomment this to deploy to a custom env name
+# aws_env="your-env-name-here"
+
 # turn CAL_BASE_IMAGE into CSYS_VER by splitting at the :, splitting again by underscore and keeping the
 # first two fields, and then converting to lowercase
 CSYS_VER=${CAL_BASE_IMAGE##*:}
@@ -19,27 +29,37 @@ CSYS_VER=`echo $CSYS_VER | awk '{print tolower($0)}'`
 # variables that will be changed less-frequently
 TMP_INSTALL_DIR="/tmp/calcloud_install"
 
-mkdir $TMP_INSTALL_DIR
-cd $TMP_INSTALL_DIR
+# setting up the calcloud source dir if it needs downloaded
+# equivalent to "if len($var) == 0"
+if [ -z "${CALCLOUD_BUILD_DIR}" ]
+then
+    mkdir -p $TMP_INSTALL_DIR
+    CALCLOUD_BUILD_DIR="${TMP_INSTALL_DIR}/calcloud-$CALCLOUD_VER"
+    # calcloud source download/unpack
+    cd $TMP_INSTALL_DIR
+    wget "https://github.com/spacetelescope/calcloud/archive/v$CALCLOUD_VER.tar.gz"
+    tar -xvzf "v$CALCLOUD_VER.tar.gz"
+    rm "v$CALCLOUD_VER.tar.gz"  
+fi
 
-# calcloud source download/unpack
-wget "https://github.com/spacetelescope/calcloud/archive/v$CALCLOUD_VER.tar.gz"
-tar -xvzf "v$CALCLOUD_VER.tar.gz"
-rm "v$CALCLOUD_VER.tar.gz"
-
-# caldp source download/unpack
-# github's tarballs don't work with pip install, so we have to clone and checkout the tag
-git clone https://github.com/spacetelescope/caldp.git
-cd caldp && git fetch --all --tags && git checkout tags/v${CALDP_VER} && cd ..
+# setting up the caldp source dir if it needs downloaded
+# equivalent to "if len($var) == 0"
+if [ -z "${CALDP_BUILD_DIR}"]
+then
+    mkdir -p $TMP_INSTALL_DIR
+    CALDP_BUILD_DIR="${TMP_INSTALL_DIR}/caldp"
+    cd $TMP_INSTALL_DIR
+    # caldp source download/unpack
+    # github's tarballs don't work with pip install, so we have to clone and checkout the tag
+    git clone https://github.com/spacetelescope/caldp.git
+    cd caldp && git fetch --all --tags && git checkout tags/v${CALDP_VER} && cd ..
+fi
 
 # get a couple of things from AWS ssm
 # the env, i.e. sb,dev,test,prod
 aws_env_response=`awsudo $ADMIN_ARN aws ssm get-parameter --name "environment" | grep "Value"`
 aws_env=${aws_env_response##*:}
 aws_env=`echo $aws_env | tr -d '",'`
-
-#uncomment this to deploy to a custom env name
-# aws_env="your-env-name-here"
 
 # the tf state bucket name
 aws_tfstate_response=`awsudo $ADMIN_ARN aws ssm get-parameter --name "/s3/tfstate" | grep "Value"`
@@ -48,7 +68,7 @@ aws_tfstate=`echo $aws_tfstate | tr -d '",'`
 echo $aws_tfstate
 
 # initial terraform setup
-cd calcloud-${CALCLOUD_VER}/terraform
+cd ${CALCLOUD_BUILD_DIR}/terraform
 
 # terraform init and s3 state backend config
 awsudo $ADMIN_ARN terraform init -backend-config="bucket=${aws_tfstate}" -backend-config="key=calcloud/${aws_env}.tfstate" -backend-config="region=us-east-1"
@@ -61,25 +81,54 @@ repo_url=${repo_url_response##*=}
 # removes double quotes from variable
 repo_url=`echo $repo_url | tr -d '"'`
 
-# build and deploy caldp docker image
-cd ../../caldp
-# cd ~/bhayden/caldp
+##### DOCKER IMAGE BUILDING #########
 CALDP_DOCKER_IMAGE="${repo_url}:${CALDP_IMAGE_TAG}"
-docker build -f Dockerfile -t ${CALDP_DOCKER_IMAGE} --build-arg CAL_BASE_IMAGE=${CAL_BASE_IMAGE}  .
-# need to "log in" to ecr to push the image
+MODEL_DOCKER_IMAGE="${repo_url}:model"
+TRAINING_DOCKER_IMAGE="${repo_url}:training"
+
+# need to "log in" to ecr to push or pull the images
 awsudo $ADMIN_ARN aws ecr get-login-password --region us-east-1 | docker login --username AWS --password-stdin $repo_url
+
+# naming is confusing here but "modeling" directory plus "training" image is correct
+cd ${CALCLOUD_BUILD_DIR}/modeling
+set -o pipefail && docker build -f Dockerfile -t "${TRAINING_DOCKER_IMAGE}" .
+training_docker_build_status=$?
+if [[ $training_docker_build_status -ne 0 ]]; then
+    echo "training job docker build failed; exiting"
+    exit 1
+fi
+
+# jobPredict lambda env
+cd ${CALCLOUD_BUILD_DIR}/lambda/JobPredict
+set -o pipefail && docker build -f Dockerfile -t "${MODEL_DOCKER_IMAGE}" .
+model_docker_build_status=$?
+if [[ $model_docker_build_status -ne 0 ]]; then
+    echo "predict lambda env docker build failed; exiting"
+    exit 1
+fi
+
+# caldp image
+cd ${CALDP_BUILD_DIR}
+set -o pipefail && docker build -f Dockerfile -t "${CALDP_DOCKER_IMAGE}" --build-arg CAL_BASE_IMAGE="${CAL_BASE_IMAGE}"  .
+caldp_docker_build_status=$?
+if [[ $caldp_docker_build_status -ne 0 ]]; then
+    echo "caldp docker build failed; exiting"
+    exit 1
+fi
+
+docker push ${TRAINING_DOCKER_IMAGE}
+docker push ${MODEL_DOCKER_IMAGE}
 docker push ${CALDP_DOCKER_IMAGE}
 
-# deploy rest of terraform
-cd ../calcloud-${CALCLOUD_VER}/terraform
+#### PRIMARY TERRAFORM BUILD #####
+cd ${CALCLOUD_BUILD_DIR}/terraform
+
 # must taint the compute env to be safe about launch template handling. see comments in batch.tf
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[0]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[1]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[2]
 awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.compute_env[3]
-
-awsudo $ADMIN_ARN terraform taint docker_registry_image.calcloud_predict_model
-awsudo $ADMIN_ARN terraform taint module.lambda_function_container_image.aws_lambda_function.this[0]
+awsudo $ADMIN_ARN terraform taint aws_batch_compute_environment.model_compute_env[0]
 
 # manual confirmation required
 awsudo $ADMIN_ARN terraform apply -var "awsysver=${CALCLOUD_VER}" -var "awsdpver=${CALDP_VER}" -var "csys_ver=${CSYS_VER}" -var "environment=${aws_env}"
